fix: refatoração para adicionar as duas funções em um arquivo
